cypress: split commands.js in several files
* Splits commands.js in several files for better readability.
* Adds circulation custom commands.
