+++ tests/e2e/cypress/cypress/support/commands.js
@@ -57,127 +57,6 @@ Cypress.Commands.add("setup", () => {
 })
 //#endregion
 
-//#region User
-// Logout
-Cypress.Commands.add("logout", () => {
-  // click on username
-  cy.get('#my-account-menu').click()
-  // then click on Logout link
-  cy.get('#logout-menu').click()
-})
-
-Cypress.Commands.add("login", (email, password) => {
-  // click on "My account"
-  cy.get('#my-account-menu').click()
-  cy.get('#login-menu').click()
-  cy.get('#email').type(email)
-  cy.get('#password').type(password)
-  cy.get('form[name="login_user_form"]').submit()
-})
-
-// Login to professional interface
-Cypress.Commands.add("adminLogin", (email, password) => {
-  cy.login(email, password)
-  cy.wait(1000)
-
-  // set language to english BEFORE going to professional interface
-  cy.setLanguageToEnglish()
-
-  // go to professional interface
-  cy.get('#my-account-menu').click()
-  cy.get('#professional-interface-menu').click()
-  cy.url().should('include', '/professional/')
- })
- //#endregion
-
- //#region Navigation
-// Go to user profile (using menus)
-// tabId: (not mandatory). If exists: go directly to given tabId
-Cypress.Commands.add("userProfile", (tabId) => {
-  // we should be already connected
-  cy.get('#my-account-menu').click()
-  cy.get('#my-profile-menu').click()
-  cy.wait(2500)
-  // Go to tabId if exist
-  if (tabId !== undefined) {
-    cy.get('#' + tabId).click()
-  }
-})
-
-// Go to a specific menu from professional homepage
-// menuId: `id=` attribute content
-Cypress.Commands.add("goToMenu", (menuId) => {
-  // Go to professional homepage
-  cy.get('#homepage-logo').click()
-  // if already on professional, do nothing
-  cy.url().then((url) => {
-    if (!url.includes('/professional/')) {
-      cy.get('#my-account-menu').click()
-      cy.get('#professional-interface-menu').click()
-      cy.wait(800)
-    }
-  })
-
-  // Check we're on admin page
-  cy.url().should('include', '/professional/')
-  // Click on 'menuTitle' from Catalog menu
-  cy.get('#' + menuId).click()
-})
-
-Cypress.Commands.add("goToItem", (itemBarcode) => {
-  // Go to homepage
-  cy.get('#homepage-logo').click()
-  cy.wait(800)
-  cy.url().then((url) => {
-    if (url.includes('/professional/')) {
-      // on professional context
-      cy.get('.form-control').type(itemBarcode).type('{enter}')
-    }
-    else {
-       // on public context
-      cy.get('.d-none > main-search-bar > .flex-grow-1 > .rero-ils-autocomplete > .form-control').type(itemBarcode).type('{enter}')
-    }
-  })
-  // Use first element
-  cy.get('.card-title > a').click()
-})
-//#endregion
-
-//#region Record
-Cypress.Commands.add("createItem", (barcode, itemType, localisation) => {
-  // Go to Catalog > Documents
-  cy.goToMenu('documents-menu-frontpage')
-  // Use one document (between first and tenth)
-  let randomInteger = Math.floor((Math.random() * 9) + 1);
-  cy.wait(2100) // because ngx-spinner hide the menu with a black transparent filter
-  cy.get(':nth-child(' + randomInteger.toString() + ') > ng-core-record-search-result > admin-documents-brief-view > .card-title > a').click()
-  // Click on "Add…" button to add an item
-  cy.get('.col > .btn').click()
-  // Fill in Item barcode
-  cy.get('#barcode').type(barcode)
-  // Wait that barcode to be checked (by API)
-  cy.wait(800)
-  // Fill in Item Call number with barcode content
-  cy.get('#call_number').type(barcode)
-  // Fill in Item Category
-  cy.get('select').first().select(itemType)
-  // Fill in localisation (could be 0, 1, etc. to select first element from select list)
-  cy.get('select').eq(1).select(localisation)
-
-  // Validate the form
-  cy.get('.mt-4 > [type="submit"]').click()
-
-  // Assert that the item has been created
-  cy.contains(barcode, {timeout: 8000})
-})
-//#endregion
-
-Cypress.Commands.add("deleteRecordFromDetailView", () => {
-  // Delete record and confirm deletion
-  cy.get('#detail-delete-button').click();
-  cy.get('#modal-confirm-button').click();
-});
-
 //#region Document editor ----------------------------------------------------->
 Cypress.Commands.add("populateSimpleRecord", (document) => {
   cy.fixture('document').then(function (documentData) {
